+++ tests/BiosTests.hs
@@ -4,6 +4,7 @@
 {-# LANGUAGE CPP #-}
 {-# LANGUAGE FlexibleInstances #-}
 {-# LANGUAGE TypeApplications #-}
+{-# LANGUAGE NumericUnderscores #-}
 module Main (main) where
 
 import Utils
@@ -17,7 +18,10 @@ import qualified Test.Tasty.Runners     as Tasty
 import HIE.Bios
 import HIE.Bios.Cradle
 import HIE.Bios.Cradle.Cabal (cabalBuildDir)
-import HIE.Bios.Types (LoadMode(..))
+import HIE.Bios.Types (LoadMode(..), CacheDir (..))
+import Control.Concurrent (threadDelay)
+import Control.Concurrent.Async (replicateConcurrently)
+import Control.Exception (SomeException, evaluate, try)
 import Control.Monad (forM_, forM, unless, when)
 import Control.Monad.Extra (unlessM)
 import Control.Monad.IO.Class
@@ -29,9 +33,11 @@ import System.Directory
 import System.FilePath ((</>), makeRelative)
 import System.Info.Extra (isWindows)
 import System.IO (BufferMode (LineBuffering), hSetBuffering, stderr, stdout)
+import System.IO.Temp
 import qualified HIE.Bios.Ghc.Gap as Gap
 import HIE.Bios.Cradle.Utils (expandGhcOptionResponseFile)
-import HIE.Bios.Environment (extractUnits)
+import HIE.Bios.Environment (extractUnits, resolveCacheDir)
+import HIE.Bios.Process (cacheFileIn)
 
 
 argDynamic :: [String]
@@ -81,6 +87,7 @@ main = do
     testGroup "Bios-tests"
       [ testGroup "Find cradle" findCradleTests
       , testGroup "Symlink" symbolicLinkTests
+      , testGroup "Cache files" cacheFileTests
       , testGroup "Loading tests"
         [ testGroup "bios" biosTestCases
         , testGroup "direct" directTestCases
@@ -134,6 +141,29 @@ symbolicLinkTests =
           assertFailure "Test invariant broken, this file must exist."
         loadComponentOptions "./c/A.hs" []
         assertLoadNone
+  ]
+
+-- | Concurrent 'cacheFileIn' calls can race on the same cache entry.
+cacheFileTests :: [TestTree]
+cacheFileTests =
+  [ testCase "cacheFile is atomic under concurrent calls" $ do
+      withSystemTempDirectory "hie-bios-cache-file-test" $ \testCacheDir -> do
+        let payloadPrefix = "#!/bin/sh\n"
+            payloadSuffix = concat (replicate 500 "echo 'cacheFile is atomic under concurrent calls'\n")
+            payload = payloadPrefix <> payloadSuffix
+            -- Pause mid-write so a non-atomic populate would expose a partial
+            -- file, and so all threads pile up on the entry concurrently.
+            populate fp = do
+              writeFile fp payloadPrefix
+              threadDelay 50_000 -- 50ms
+              appendFile fp payloadSuffix
+        results <- replicateConcurrently 16 $ try @SomeException $ do
+          fp <- cacheFileIn (CacheDir testCacheDir) "ghc-pkg" "0123456789abcdef" populate
+          contents <- readFile fp
+          contents <$ evaluate (length contents)
+        forM_ results $ \case
+          Left err -> assertFailure $ "cacheFile threw: " <> show err
+          Right contents -> assertEqual "cached file contents" payload contents
   ]
 
 biosTestCases :: [TestTree]
@@ -211,7 +241,9 @@ cabalTestCases extraGhcDep =
       initCradle "B.hs"
       assertCradle isCabalCradle
       root <- askRoot
-      buildDir <- liftIO $ cabalBuildDir root
+      buildDir <- liftIO $ do
+        cacheDir <- resolveCacheDir "" Nothing
+        cabalBuildDir cacheDir root
       -- use --multi-repl, as that was the codepath with the bug
       loadFileGhc "B.hs" []
       liftIO $ do
@@ -221,6 +253,16 @@ cabalTestCases extraGhcDep =
         -- Check we are using the correct build directory
         buildDirExists <- doesDirectoryExist buildDir
         assertBool "build dir does not exist" buildDirExists
+  , biosTestCase "custom cache dir" $ runTestEnv "./simple-cabal" $
+      withSystemTempDirectory "hie-bios-custom-cache-dir" $ \cacheRoot -> do
+        initCradleWithConfig defaultCradleRunConfig { cradleCacheDir = Just (CacheDir cacheRoot) } "B.hs"
+        assertCradle isCabalCradle
+        loadComponentOptions "B.hs" []
+        _ <- assertLoadSuccess
+        liftIO $ do
+          entries <- listDirectory cacheRoot
+          assertBool ("cache entries under the configured dir: " <> show entries)
+            (any (\e -> "wrapper" `isInfixOf` e || "dist-" `isPrefixOf` e) entries)
   , biosTestCaseAll "nested-cabal" $ runTestEnv "./nested-cabal" $ do
       attemptCabalSingleTargetLoad "sub-comp/Lib.hs"
       mode <- askLoadMode