+++ ChangeLog.md
@@ -1,5 +1,14 @@
 # ChangeLog hie-bios
 
+## Unreleased
+
+* Make cacheFile thread-safe using file locks ([#520](https://github.com/haskell/hie-bios/pull/520)).
+* Make functions take the used cache directory explicitly ([#520](https://github.com/haskell/hie-bios/pull/520)).
+  * Add *WithConfig functions.
+  * Adjust existing functions to take a cache directory.
+  * Replace thread-unsafety workaround of initSession' and
+    initSessionWithMessage' with the cache directory argument.
+
 ## 2026-06-30 - 0.20.0
 
 * Hide LoadMode in TestM ([#516](https://github.com/haskell/hie-bios/pull/516))